+++ src/Spago/Dhall.hs
@@ -22,9 +22,12 @@ type DhallExpr a = Dhall.Expr Parser.Src a
 
 
 -- | Format a Dhall file in ASCII
-format :: MonadIO m => Text -> m ()
-format pathText = liftIO $ Dhall.Format.format
-  (Dhall.Format.Format Dhall.Pretty.ASCII $ Dhall.Format.Modify (Just $ Text.unpack pathText))
+format :: MonadIO m => DoFormat -> Text -> m ()
+format shouldFormat pathText =
+  when (shouldFormat == DoFormat) $
+    liftIO $ Dhall.Format.format
+      (Dhall.Format.Format Dhall.Pretty.ASCII $
+       Dhall.Format.Modify (Just $ Text.unpack pathText))
 
 
 -- | Prettyprint a Dhall expression
@@ -50,15 +53,14 @@ readRawExpr pathText = do
     else (pure Nothing)
 
 
-writeRawExpr :: Bool -> Text -> (Text, DhallExpr Dhall.Import) -> IO ()
-writeRawExpr hasDontFormatConfig pathText (header, expr) = do
+writeRawExpr :: DoFormat -> Text -> (Text, DhallExpr Dhall.Import) -> IO ()
+writeRawExpr shouldFormat pathText (header, expr) = do
   -- After modifying the expression, we have to check if it still typechecks
   -- if it doesn't we don't write to file.
   resolvedExpr <- Dhall.Import.load expr
   _ <- throws (Dhall.TypeCheck.typeOf resolvedExpr)
-  unless hasDontFormatConfig $ do
-    writeTextFile (pathFromText pathText) $ prettyWithHeader header expr <> "\n"
-    format pathText
+  writeTextFile (pathFromText pathText) $ prettyWithHeader header expr <> "\n"
+  format shouldFormat pathText
 
 
 -- | Returns a Dhall Text literal from a lone string