+++ src/Spago/Prelude.hs
@@ -9,6 +9,7 @@ module Spago.Prelude
   , pathFromText
   , assertDirectory
   , GlobalOptions (..)
+  , DoFormat (..)
   , Spago
   , module X
   , Typeable
@@ -114,9 +115,12 @@ instance Show SpagoError where
   show (SpagoError err) = Text.unpack err
 
 
+-- | Flag to skip automatic formatting of the Dhall files
+data DoFormat = DoFormat | NoFormat deriving (Eq)
+
 data GlobalOptions = GlobalOptions
   { debug :: Bool
-  , dontFormatConfig :: Bool
+  , shouldFormat :: DoFormat
   }
 
 type Spago m =